Things, that didn't made to 2017.7 for some reasons

+++ tests/integration/modules/test_beacons.py
@@ -58,7 +58,7 @@ class BeaconsTest(ModuleCase):
 
     @classmethod
     def tearDownClass(cls):
-        if os.path.isfile(cls.beacons_config_file_path):
+        if cls.beacons_config_file_path and os.path.isfile(cls.beacons_config_file_path):
             os.unlink(cls.beacons_config_file_path)
 
     def setUp(self):
